+++ anyblok_marshmallow/__init__.py
@@ -9,4 +9,4 @@
 
 from .schema import ModelSchema, PostLoadSchema  # noqa
 from .exceptions import RegistryNotFound  # noqa
-from .fields import Nested, File, Text, JsonCollection  # noqa
+from .fields import Nested, File, Text, JsonCollection, PhoneNumber  # noqa